Move test_scheduler_waited_by_drop_bank_service
